Finally configure wobble on max LFO
Not the most satisfying implementationm, feels like there must be a way of implementing it without using as much resources. We have to calculate frequency twice for the stereo channels in order for the wobble to be independent, but then we need to take the lfo phase increase out of the equation if we want it to run at the same speed as the monophonic algos.